+++ src/com/jpexs/decompiler/flash/gui/generictageditors/UUIDEditor.java
@@ -161,4 +161,9 @@ public class UUIDEditor extends JTextField implements GenericTagEditor {
     @Override
     public void validateValue() {
     }
+    
+    @Override
+    public Object getObject() {
+        return obj;
+    }
 }