+++ CHANGELOG.md
@@ -29,6 +29,7 @@ All notable changes to this project will be documented in this file.
 - AS1/2 new keyword on empty method name
 - AS2 getters and setters decoding
 - [#116] §§push at the end of switch branches
+- Convolution matrix filter display and editing
 
 ## [19.0.0] - 2023-10-01
 ### Added